refactor(build): 域构建入口归位 src/build.js 与 src-cordis/build.js，preset 层留 src-cordis/build/
布局定稿（轻域化，非 monorepo——版本/发布/lockfile 仍单一）：src 域构建入口
src/build.js（与 rspack.config.mjs 同根，配置随源码）；src-cordis 域入口
src-cordis/build.js（preset 层留 src-cordis/build/{service,client}-config.mjs +
每包 cordis.config.mjs 描述）；根 scripts/ 只留共享（build-common/loader/流程脚本：
pack/bump/syncver/postbump/version-common）。build.mjs 聚合文件删除，编排全在
package.json（build = build:src && build:cordis）。template-loader 随 src 域移
src/。共享 build-common.mjs 抽 collect/URL 回写/terser/assert 供两域复用。
全链验证 build（src+cordis）通过、pack/syncver 一致。
